breaking: remove ClassAccessorProperty from Standardized
