+++ packages/babel-types/src/definitions/core.ts
@@ -2214,83 +2214,6 @@ defineType("ClassProperty", {
   },
 });
 
-defineType("ClassAccessorProperty", {
-  visitor: ["decorators", "key", "typeAnnotation", "value"],
-  builder: [
-    "key",
-    "value",
-    "typeAnnotation",
-    "decorators",
-    "computed",
-    "static",
-  ],
-  aliases: ["Property", "Accessor"],
-  ...classMethodOrPropertyUnionShapeCommon(true),
-  fields: {
-    ...classMethodOrPropertyCommon(),
-    key: {
-      validate: chain(
-        (function () {
-          const normal = assertNodeType(
-            "Identifier",
-            "StringLiteral",
-            "NumericLiteral",
-            "BigIntLiteral",
-            "PrivateName",
-          );
-          const computed = assertNodeType("Expression");
-
-          return function (
-            node: t.ClassAccessorProperty,
-            key: string,
-            val: any,
-          ) {
-            const validator = node.computed ? computed : normal;
-            validator(node, key, val);
-          } satisfies ValidatorImpl;
-        })(),
-        assertNodeType(
-          "Identifier",
-          "StringLiteral",
-          "NumericLiteral",
-          "BigIntLiteral",
-          "Expression",
-          "PrivateName",
-        ),
-      ),
-    },
-    value: {
-      validate: assertNodeType("Expression"),
-      optional: true,
-    },
-    definite: {
-      validate: assertValueType("boolean"),
-      optional: true,
-    },
-    typeAnnotation: {
-      validate: assertNodeType("TypeAnnotation", "TSTypeAnnotation"),
-
-      optional: true,
-    },
-    decorators: {
-      validate: arrayOfType("Decorator"),
-      optional: true,
-    },
-    readonly: {
-      validate: assertValueType("boolean"),
-      optional: true,
-    },
-    declare: {
-      validate: assertValueType("boolean"),
-      optional: true,
-    },
-    variance: {
-      validate: assertNodeType("Variance"),
-      optional: true,
-    },
-  },
-});
-
 defineType("ClassPrivateProperty", {
   visitor: ["decorators", "variance", "key", "typeAnnotation", "value"],
   builder: ["key", "value", "decorators", "static"],
