+++ packages/babel-types/src/definitions/experimental.ts
@@ -1,4 +1,15 @@
-import defineType, { assertNodeType, assertValueType } from "./utils.ts";
+import {
+  classMethodOrPropertyUnionShapeCommon,
+  classMethodOrPropertyCommon,
+} from "./core.ts";
+import type * as t from "../index.ts";
+import defineType, {
+  arrayOfType,
+  assertNodeType,
+  assertValueType,
+  chain,
+  type ValidatorImpl,
+} from "./utils.ts";
 
 // https://github.com/tc39/proposal-partial-application
 defineType("ArgumentPlaceholder", {});
@@ -18,6 +29,83 @@ defineType("BindExpression", {
 });
 
 // https://github.com/tc39/proposal-decorators
+defineType("ClassAccessorProperty", {
+  visitor: ["decorators", "key", "typeAnnotation", "value"],
+  builder: [
+    "key",
+    "value",
+    "typeAnnotation",
+    "decorators",
+    "computed",
+    "static",
+  ],
+  aliases: ["Property", "Accessor"],
+  ...classMethodOrPropertyUnionShapeCommon(true),
+  fields: {
+    ...classMethodOrPropertyCommon(),
+    key: {
+      validate: chain(
+        (function () {
+          const normal = assertNodeType(
+            "Identifier",
+            "StringLiteral",
+            "NumericLiteral",
+            "BigIntLiteral",
+            "PrivateName",
+          );
+          const computed = assertNodeType("Expression");
+
+          return function (
+            node: t.ClassAccessorProperty,
+            key: string,
+            val: any,
+          ) {
+            const validator = node.computed ? computed : normal;
+            validator(node, key, val);
+          } satisfies ValidatorImpl;
+        })(),
+        assertNodeType(
+          "Identifier",
+          "StringLiteral",
+          "NumericLiteral",
+          "BigIntLiteral",
+          "Expression",
+          "PrivateName",
+        ),
+      ),
+    },
+    value: {
+      validate: assertNodeType("Expression"),
+      optional: true,
+    },
+    definite: {
+      validate: assertValueType("boolean"),
+      optional: true,
+    },
+    typeAnnotation: {
+      validate: assertNodeType("TypeAnnotation", "TSTypeAnnotation"),
+
+      optional: true,
+    },
+    decorators: {
+      validate: arrayOfType("Decorator"),
+      optional: true,
+    },
+    readonly: {
+      validate: assertValueType("boolean"),
+      optional: true,
+    },
+    declare: {
+      validate: assertValueType("boolean"),
+      optional: true,
+    },
+    variance: {
+      validate: assertNodeType("Variance"),
+      optional: true,
+    },
+  },
+});
+
 defineType("Decorator", {
   visitor: ["expression"],
   fields: {
